Add example of One Definition Rule violation with "private" requires
This example attempts to illustrate a real world package my team maintains.

SharedStaticLib represents a collection of helpful utilities used by a variety
of tools. SharedStaticLib is careful to minimize its public API footprint and keeps
any heavy weight dependencies out of its public headers / API.

Use of libraries like Boost or zlib are seen as "implementation details" of the SharedStaticLib
and otherwise hidden from consuming apps.

Other tools, like MyApp, consume SharedStaticLib and may also have some of these same
dependencies (e.g. Boost or zlib) for their own reasons.

This example shows that using "private" requirements in a shared library
may result in violations of the One Definition Rule (ODR).

The easiest one to see is Boost, where this example shows
that both Boost 1.70.0 and Boost 1.67.0 can end up in the same binary

Conan is not honoring the full_version_mode requirement of Boost by the SharedStaticLib.

Likewise, with zlib, the consuming app has selected a higher, semver compatible
version of zlib, but Conan is not factoring this into its decision to use
the cached SharedStaticLib, which has an older version embedded inside it,
which would similarly result in ODR violations because the static lib
is using 1.2.8 and the consuming app 1.2.11 yet both are found in the consuming app.